pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uav-edge-ai-stack"
version = "0.1.0"
description = "Industry-style MAVLink UAV autonomy stack scaffold"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "UAV Project Team" }]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "autonomy/src"}

[tool.setuptools.packages.find]
where = ["autonomy/src"]

[tool.pytest.ini_options]
pythonpath = ["autonomy/src", "."]
testpaths = ["autonomy/tests"]
addopts = "-q"

[tool.black]
line-length = 100
